refactor(adapter): take `space.ts` apart
