+++ glass-easel-miniprogram-adapter/src/builder/index.ts
@@ -1,0 +1,2 @@
+export { BehaviorBuilder } from './behavior_builder'
+export { ComponentBuilder } from './component_builder'
